GEOMESA-880 Updating surefire plugin to 2.18.1
* Modified some unit tests that were broken
* Split AccumuloDataStoreTest up into smaller parts
