+++ geomesa-jobs/pom.xml
@@ -139,7 +139,7 @@
                     <plugin>
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.17</version>
+                        <version>2.18.1</version>
                         <configuration>
                             <!-- the shaded feature jar causes test issues -->
                             <skipTests>true</skipTests>
