+++ geomesa-kafka/geomesa-kafka-datastore/src/test/scala/org/locationtech/geomesa/kafka/consumer/offsets/OffsetManagerIntegrationTest.scala
@@ -25,48 +25,48 @@ class OffsetManagerIntegrationTest extends Specification with HasEmbeddedKafka {
 
   sequential // this doesn't really need to be sequential, but we're trying to reduce zk load
 
+  val props = new Properties
+  props.put("group.id", "mygroup")
+  props.put("metadata.broker.list", brokerConnect)
+  props.put("zookeeper.connect", zkConnect)
+  val config = new ConsumerConfig(props)
+  val offsetManager = new OffsetManager(config)
+
+  val topic = "test"
+
+  step {
+    val producerProps = new Properties()
+    producerProps.put("metadata.broker.list", brokerConnect)
+    producerProps.put("serializer.class", "kafka.serializer.DefaultEncoder")
+    val producer = new Producer[Array[Byte], Array[Byte]](new ProducerConfig(producerProps))
+    for (i <- 0 until 10) {
+      producer.send(new KeyedMessage(topic, i.toString.getBytes("UTF-8"), s"test $i".getBytes("UTF-8")))
+    }
+    producer.close()
+  }
+
   "OffsetManager" should {
-    "find offsets" >> {
-      val props = new Properties
-      props.put("group.id", "mygroup")
-      props.put("metadata.broker.list", brokerConnect)
-      props.put("zookeeper.connect", zkConnect)
-      val config = new ConsumerConfig(props)
-      val offsetManager = new OffsetManager(config)
-
-      val topic = "test"
-
-      val producerProps = new Properties()
-      producerProps.put("metadata.broker.list", brokerConnect)
-      producerProps.put("serializer.class", "kafka.serializer.DefaultEncoder")
-      val producer = new Producer[Array[Byte], Array[Byte]](new ProducerConfig(producerProps))
-      for (i <- 0 until 10) {
-        producer.send(new KeyedMessage(topic, i.toString.getBytes("UTF-8"), s"test $i".getBytes("UTF-8")))
-      }
-      producer.close()
-
-      "by number" >> {
-        offsetManager.getOffsets(topic, SpecificOffset(1)) mustEqual Map(TopicAndPartition(topic, 0) -> 1)
-      }
-      "by earliest" >> {
-        offsetManager.getOffsets(topic, EarliestOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 0)
-      }
-      "by latest" >> {
-        offsetManager.getOffsets(topic, LatestOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 10)
-      }
-      "by group" >> {
-        offsetManager.commitOffsets(Map(TopicAndPartition(topic, 0) -> OffsetAndMetadata(5)))
-        offsetManager.getOffsets(topic, GroupOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 5)
-      }
-      "by binary search" >> {
-        val decoder = new StringDecoder()
-        val offset = FindOffset((m: Message) => {
-          val bb = Array.ofDim[Byte](m.payload.remaining())
-          m.payload.get(bb)
-          decoder.fromBytes(bb).substring(5).toInt.compareTo(7)
-        })
-        offsetManager.getOffsets(topic, offset) mustEqual Map(TopicAndPartition(topic, 0) -> 7)
-      }
+    "find offsets by number" in {
+      offsetManager.getOffsets(topic, SpecificOffset(1)) mustEqual Map(TopicAndPartition(topic, 0) -> 1)
+    }
+    "find offsets by earliest" in {
+      offsetManager.getOffsets(topic, EarliestOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 0)
+    }
+    "find offsets by latest" in {
+      offsetManager.getOffsets(topic, LatestOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 10)
+    }
+    "find offsets by group" in {
+      offsetManager.commitOffsets(Map(TopicAndPartition(topic, 0) -> OffsetAndMetadata(5)))
+      offsetManager.getOffsets(topic, GroupOffset) mustEqual Map(TopicAndPartition(topic, 0) -> 5)
+    }
+    "find offsets by binary search" in {
+      val decoder = new StringDecoder()
+      val offset = FindOffset((m: Message) => {
+        val bb = Array.ofDim[Byte](m.payload.remaining())
+        m.payload.get(bb)
+        decoder.fromBytes(bb).substring(5).toInt.compareTo(7)
+      })
+      offsetManager.getOffsets(topic, offset) mustEqual Map(TopicAndPartition(topic, 0) -> 7)
     }
   }
