+++ geomesa-stream/geomesa-stream-datastore/src/test/scala/org/locationtech/geomesa/stream/datastore/StreamDataStoreTest.scala
@@ -29,96 +29,97 @@ class StreamDataStoreTest extends Specification {
 
   sequential
 
+  val count  = new AtomicLong(0)
+  val count2 = new AtomicLong(0)
+  val count3 = new AtomicLong(0)
+
+  val ff = CommonFactoryFinder.getFilterFactory2()
+  val sourceConf =
+    """
+      |{
+      |  type         = "generic"
+      |  source-route = "netty4:tcp://localhost:5899?textline=true"
+      |  sft          = {
+      |                   type-name = "testdata"
+      |                   fields = [
+      |                     { name = "label",     type = "String" }
+      |                     { name = "geom",      type = "Point",  index = true, srid = 4326, default = true }
+      |                     { name = "dtg",       type = "Date",   index = true }
+      |                   ]
+      |                 }
+      |  converter    = {
+      |                   id-field = "md5(string2bytes($0))"
+      |                   type = "delimited-text"
+      |                   format = "DEFAULT"
+      |                   fields = [
+      |                     { name = "label",     transform = "trim($1)" }
+      |                     { name = "geom",      transform = "point($2::double, $3::double)" }
+      |                     { name = "dtg",       transform = "datetime($4)" }
+      |                   ]
+      |                 }
+      |}
+    """.stripMargin
+
+  val sds = DataStoreFinder.getDataStore(Map(
+      StreamDataStoreParams.STREAM_DATASTORE_CONFIG.key -> sourceConf,
+      StreamDataStoreParams.CACHE_TIMEOUT.key -> Integer.valueOf(2)
+    )).asInstanceOf[StreamDataStore]
+
   "StreamDataStore" should {
-
-    val ff = CommonFactoryFinder.getFilterFactory2()
-    val sourceConf =
-      """
-        |{
-        |  type         = "generic"
-        |  source-route = "netty4:tcp://localhost:5899?textline=true"
-        |  sft          = {
-        |                   type-name = "testdata"
-        |                   fields = [
-        |                     { name = "label",     type = "String" }
-        |                     { name = "geom",      type = "Point",  index = true, srid = 4326, default = true }
-        |                     { name = "dtg",       type = "Date",   index = true }
-        |                   ]
-        |                 }
-        |  converter    = {
-        |                   id-field = "md5(string2bytes($0))"
-        |                   type = "delimited-text"
-        |                   format = "DEFAULT"
-        |                   fields = [
-        |                     { name = "label",     transform = "trim($1)" }
-        |                     { name = "geom",      transform = "point($2::double, $3::double)" }
-        |                     { name = "dtg",       transform = "datetime($4)" }
-        |                   ]
-        |                 }
-        |}
-      """.stripMargin
-
-    val ds = DataStoreFinder.getDataStore(
-      Map(
-        StreamDataStoreParams.STREAM_DATASTORE_CONFIG.key -> sourceConf,
-        StreamDataStoreParams.CACHE_TIMEOUT.key -> Integer.valueOf(2)
-      ))
-
     "be built from a conf string" >> {
-      ds must not be null
+      sds must not beNull
     }
 
-    val count = new AtomicLong(0)
-    val listener = new StreamListener {
-      override def onNext(sf: SimpleFeature): Unit = {
-        count.incrementAndGet()
+    "read and write" >> {
+      val listener = new StreamListener {
+        override def onNext(sf: SimpleFeature): Unit = {
+          count.incrementAndGet()
+        }
       }
-    }
-    val sds = ds.asInstanceOf[StreamDataStore]
-    sds.registerListener(listener)
+      sds.registerListener(listener)
 
-    val count2 = new AtomicLong(0)
-    val listener2 = StreamListener { sf => count2.incrementAndGet() }
-    sds.registerListener(listener2)
+      val listener2 = StreamListener { sf => count2.incrementAndGet() }
+      sds.registerListener(listener2)
 
-    val count3 = new AtomicLong(0)
-    val bboxFilter = ff.bbox("geom", 49.0, 79.0, 51.0, 80.0, "EPSG:4326")
-    val listener3 = StreamListener(bboxFilter, _ =>  count3.incrementAndGet())
-    sds.registerListener(listener3)
+      val bboxFilter = ff.bbox("geom", 49.0, 79.0, 51.0, 80.0, "EPSG:4326")
+      val listener3 = StreamListener(bboxFilter, _ =>  count3.incrementAndGet())
+      sds.registerListener(listener3)
 
-    val fs = ds.getFeatureSource("testdata")
+      val fs = sds.getFeatureSource("testdata")
 
-    "handle new data" >> {
-      val url = Resources.getResource("testdata.tsv")
-      val lines = Resources.readLines(url, StandardCharsets.UTF_8)
-      val socketFactory = new DefaultSocketFactory
-      Future {
-        val socket = socketFactory.createSocket("localhost", 5899)
-        val os = socket.getOutputStream
-        IOUtils.writeLines(lines, IOUtils.LINE_SEPARATOR_UNIX, os)
-        os.flush()
-        // wait for data to arrive at the server
-        Thread.sleep(4000)
-        os.close()
+      "handle new data" >> {
+        val url = Resources.getResource("testdata.tsv")
+        val lines = Resources.readLines(url, StandardCharsets.UTF_8)
+        val socketFactory = new DefaultSocketFactory
+        Future {
+          val socket = socketFactory.createSocket("localhost", 5899)
+          val os = socket.getOutputStream
+          IOUtils.writeLines(lines, IOUtils.LINE_SEPARATOR_UNIX, os)
+          os.flush()
+          // wait for data to arrive at the server
+          Thread.sleep(4000)
+          os.close()
+        }
+
+        Thread.sleep(1000)
+        fs.getFeatures(Filter.INCLUDE).features().hasNext must beTrue
       }
 
-      Thread.sleep(1000)
-      fs.getFeatures(Filter.INCLUDE).features().hasNext must beTrue
-    }
+      "support listeners" >> {
+        count.get() must equalTo(7)
+        count2.get() must equalTo(7)
+        count3.get() must equalTo(3)
+      }
 
-    "support listeners" >> {
-      count.get() must equalTo(7)
-      count2.get() must equalTo(7)
-      count3.get() must equalTo(3)
-    }
+      "handle bbox filters" >> {
+        fs.getFeatures(bboxFilter).size() must be equalTo 3
+      }
 
-    "handle bbox filters" >> {
-      fs.getFeatures(bboxFilter).size() must be equalTo 3
-    }
-
-    "expire data after the appropriate amount of time" >> {
-      Thread.sleep(3000)
-      fs.getFeatures(Filter.INCLUDE).features().hasNext must beFalse
+      "expire data after the appropriate amount of time" >> {
+        Thread.sleep(3000)
+        fs.getFeatures(Filter.INCLUDE).features().hasNext must beFalse
+      }
+      ok
     }
   }
 }
